Revert "#163655283 API documention"
