+++ README.md
@@ -1,13 +1,6 @@
 [![Build Status](https://travis-ci.com/femitj/Politico.svg?branch=develop)](https://travis-ci.com/femitj/Politico)
 [![Coverage Status](https://coveralls.io/repos/github/femitj/Politico/badge.svg?branch=develop)](https://coveralls.io/github/femitj/Politico?branch=develop)
-[![Maintainability](https://api.codeclimate.com/v1/badges/3df9afa37bc284eac41c/maintainability)](https://codeclimate.com/github/femitj/Politico/maintainability)
+[![Test Coverage](https://api.codeclimate.com/v1/badges/3df9afa37bc284eac41c/test_coverage)](https://codeclimate.com/github/femitj/Politico/test_coverage)
 
 # Politico
 Politico is a simple RESTFUL application that enables citizens give their mandate to politicians running for different government offices while building trust in the process through transparency.
-
-
-
-
-## Author
-
-Tijani Oluwafemi
