+++ package.json
@@ -41,8 +41,7 @@
     "gh-pages": "^2.0.1",
     "jsonwebtoken": "^8.4.0",
     "make-runnable": "^1.3.6",
-    "pg": "^7.8.0",
-    "swagger-ui-express": "^4.0.2"
+    "pg": "^7.8.0"
   },
   "devDependencies": {
     "babel-preset-env": "^1.7.0",
